functions/processor/rsi/requirements.txt
functions-framework==3.*
pandas==1.3.5
fsspec==2022.5.0
gcsfs==2022.5.0
google-cloud-storage==2.5.0
requests==2.22.0
pandas-ta==0.3.14b0